refactor(runtime): rename provider protocol package
Clarify that the shared Python code hosts model-owned runtime providers rather than implementing an optimized-runtime adapter.

Keep bundle schemas and the private C++ provider ABI unchanged.
