+++ python/tensorrt_model_connect/runtime_provider/__init__.py
@@ -1,7 +1,7 @@
 # SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 # SPDX-License-Identifier: Apache-2.0
 
-"""Private implementation machinery for model-owned optimized runtimes.
+"""Private provider protocol for model-owned optimized runtimes.
 
 Nothing in this package is part of the supported Model Connect Python API.
 Integrations communicate through capsule manifests and subprocess contracts,
